components/cspr_summarizers/requirements.txt
Django==4.1.6
numpy==1.24.2
pandas==1.5.3
psycopg2-binary==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
schedule
